Refactor big utils file into smaller helper scripts
