+++ packages/nhsNotifyLambda/tests/testHelpers.ts
@@ -5,7 +5,7 @@ import * as sqs from "@aws-sdk/client-sqs"
 import {v4} from "uuid"
 
 import {PSUDataItem} from "@PrescriptionStatusUpdate_common/commonTypes"
-import {NotifyDataItemMessage} from "../src/utils"
+import {NotifyDataItemMessage} from "../src/types"
 
 // Similarly mock the SQS client
 export function mockSQSClient() {
