+++ packages/nhsNotifyLambda/tests/testUtils.test.ts
@@ -37,14 +37,9 @@ jest.unstable_mockModule(
   })
 )
 
-const {
-  addPrescriptionMessagesToNotificationStateStore,
-  removeSQSMessages,
-  checkCooldownForUpdate,
-  reportQueueStatus,
-  drainQueue,
-  makeBatchNotifyRequest
-} = await import("../src/utils")
+const {checkCooldownForUpdate, addPrescriptionMessagesToNotificationStateStore} = await import("../src/dynamo")
+const {removeSQSMessages, drainQueue, reportQueueStatus} = await import("../src/sqs")
+const {makeBatchNotifyRequest} = await import("../src/notify")
 
 const ORIGINAL_ENV = {...process.env}
 
@@ -151,7 +146,7 @@ describe("NHS notify lambda helper functions", () => {
 
     it("Throws an error if the SQS URL is not configured", async () => {
       delete process.env.NHS_NOTIFY_PRESCRIPTIONS_SQS_QUEUE_URL
-      const {drainQueue: dq} = await import("../src/utils")
+      const {drainQueue: dq} = await import("../src/sqs")
       await expect(dq(logger)).rejects.toThrow(
         "NHS_NOTIFY_PRESCRIPTIONS_SQS_QUEUE_URL not set"
       )
@@ -248,7 +243,7 @@ describe("NHS notify lambda helper functions", () => {
 
     it("Throws an error if the SQS URL is not configured", async () => {
       delete process.env.NHS_NOTIFY_PRESCRIPTIONS_SQS_QUEUE_URL
-      const {removeSQSMessages: clearFunc} = await import("../src/utils")
+      const {removeSQSMessages: clearFunc} = await import("../src/sqs")
 
       await expect(clearFunc(logger, [])).rejects.toThrow("NHS_NOTIFY_PRESCRIPTIONS_SQS_QUEUE_URL not set")
       expect(errorSpy).toHaveBeenCalledWith("Notifications SQS URL not configured")
@@ -275,7 +270,7 @@ describe("NHS notify lambda helper functions", () => {
 
     it("throws and logs error if TABLE_NAME is not set", async () => {
       delete process.env.TABLE_NAME
-      const {addPrescriptionMessagesToNotificationStateStore: addFn} = await import("../src/utils")
+      const {addPrescriptionMessagesToNotificationStateStore: addFn} = await import("../src/dynamo")
 
       await expect(
         addFn(logger, [constructPSUDataItemMessage()])
@@ -365,7 +360,7 @@ describe("NHS notify lambda helper functions", () => {
 
     it("throws if TABLE_NAME is not set", async () => {
       delete process.env.TABLE_NAME
-      const {checkCooldownForUpdate: fn} = await import("../src/utils")
+      const {checkCooldownForUpdate: fn} = await import("../src/dynamo")
       const update = constructPSUDataItemMessage().PSUDataItem
 
       await expect(fn(logger, update)).rejects.toThrow("TABLE_NAME not set")
@@ -716,7 +711,7 @@ describe("NHS notify lambda helper functions", () => {
           [process.env.MAKE_REAL_NOTIFY_REQUESTS_PARAM!]: "false"
         }
       ))
-      const {makeBatchNotifyRequest: fn} = await import("../src/utils")
+      const {makeBatchNotifyRequest: fn} = await import("../src/notify")
 
       const data = [
         constructPSUDataItemMessage({
@@ -818,7 +813,7 @@ describe("NHS notify lambda helper functions", () => {
 
     it("throws if the SQS URL is not configured", async () => {
       delete process.env.NHS_NOTIFY_PRESCRIPTIONS_SQS_QUEUE_URL
-      const {reportQueueStatus: rqs} = await import("../src/utils")
+      const {reportQueueStatus: rqs} = await import("../src/sqs")
 
       await expect(rqs(logger)).rejects.toThrow(
         "NHS_NOTIFY_PRESCRIPTIONS_SQS_QUEUE_URL not set"
